python/requirements.txt
pyserial
pynput
keyboard